REQ[#59626]: 【陈家桥】设备，测点信息完整同步至上级系统
[Description]: 【陈家桥】设备，测点信息完整同步至上级系统
[Test Suggestion]:

+++ iot-center-accessrobot/src/main/java/com/yjh/accessrobot/common/utils/ValueUtil.java
@@ -243,6 +243,6 @@ public class ValueUtil {
         return StringUtils.isEmpty(value.toString()) ? null : Integer.parseInt(value.toString());
     }
     public static BigDecimal objToBigDecimal(Object value){
-        return StringUtils.isEmpty(value.toString()) ? null : BigDecimal.valueOf(Long.parseLong(value.toString()));
+        return StringUtils.isEmpty(value.toString()) ? null : BigDecimal.valueOf(Double.parseDouble(value.toString()));
     }
 }
